[heap] Don't delay incremental marking due to concurrent sweeping
This CL aligns MinorMS with Scavenger in regards to triggering full
GC incremental marking. Incremental marking will be triggered either
after a minor GC or from a old gen allocation.

If concurrent sweeping for MinorMS is still active, it will be force
finalized. This may result in large complete sweeping phases on the
main thread. This CL does not contain a mitigation for it since it
does not seem to affect benchmarks.

Bug: v8:12612
Change-Id: I5b20473416ae31a849b05c4822cd9518be9a387f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5339887
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92684}