[wasm] Unhandlify SetWasmInternalFunction
When introducing the WasmInternalFunctionWrapper we don't always have
the value in a handle, and since the method does not allocate a handle
is not needed. Hence unhandlify the method.

Also move {WasmInternalFunction::FromExternal} to
{WasmExternalFunction::internal} and unhandlify it; this is only a
simple getter after all. This requires also handling C-API functions as
WasmExternalFunction.
This fits nicely the often-used pattern where we first check
"IsWasmExternalFunction" and then get the corresponding internal object.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: I219bfb8bb5d321b105d9e8a43559ba7c3ba3288d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5345770
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92702}