[heap] Avoid GC upon isolate deinitialization
At isolate deinitialization, the optimizing compile dispatcher is
stopped, flushing its queue and waiting for any pending compile tasks
to complete. In this process, a safepoint may be requested, which may
trigger a GC. This CL ensures that, in this case, local GC requests
will be ignored and no GC will be triggered.

Change-Id: I1426d75e093314e05b1d265e74be3b95e0355753
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5355974
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92760}