[wasm-compile-fuzzer] Modularise the fuzzer PART 1
This change templatizes the GenerateRandomWasmModule() function and the
WasmGenerator class with the WasmModuleGenerationOptions enum.

The WasmModuleGenerationOptions will be used in the upcoming changes to
specify which subset of expressions should be generated (e.g. SIMD,
wasmGC..).

This CL should not change the modules generated by the Wasm compile
fuzzer so far.
It only prepares the code to modularise the fuzzer in upcoming changes.

Change-Id: I2d5922ea4437bfe865518f17c41d8a1611f515f2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5353507
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92768}