[turboshaft][test] Add test-select-combine (disabled)
This CL just copy-pastes Turbofan test-select-combine without any
changes beyond disabling them with `#if 0`.

This looks like a fairly useless commit, but now we can port
test-branch-combine to Turboshaft, and reviewing will be much easier
(because we'll see the delta from Tubofan to Turboshaft, which we
wouldn't see easily if we were to port this file in the same CL that
creates it).

Bug: v8:12783
Change-Id: I6f183fe927f068488f04a2a5fd1b20aae2c857b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5361776
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92773}