[wasm] Move WasmApiFunctionRef to trusted space
This makes the WasmApiFunctionRef a trusted object, which allows to
reference either the WasmApiFunctionRef or the WasmTrustedInstanceData
directly as the "ref" stored in the WasmDispatchTable. We were storing a
WasmInstanceObject (or WasmApiFunctionRef) there before, which required
an additional load for the trusted data from the instance on every
indirect call to a Wasm function.

R=ahaas@chromium.org
CC=saelo@chromium.org

Bug: v8:14564
Change-Id: I32d3846abf69418ac6dd90ec0eb1cc6498c5ac13
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5355862
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92772}