[compiler][test] Remove support for Turbofan Graphs in FunctionTester
... in order to support Turboshaft with the FunctionTester.

The FunctionTester is always used with Javascript code (as a string)
as input, abd this path with a Graph as input was never used. Since
are currently porting cctests to Turboshaft, it makes more sense to
remove the support for Turbofan graphs as input, rather than add
support for Turboshaft graphs (since it would never be useful).

Bug: v8:12783
Change-Id: Ie49795d61dcd42d11e9abe1a861807a666aacbc9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5361778
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92775}