[sandbox] Tag entries in the TrustedPointerTable
Until now, we were only using the IndirectPointerTags to decide which
table to use (TrustedPointerTable or CodePointerTable), but weren't
actually tagging anything with them. With this CL, the entries in the
TPT now contain the tag in their top bits and need to be untagged with
the same tag on access. This ensures type-safety for trusted objects.

Bug: chromium:40069826
Change-Id: Ib12a01470bb0f11eacda3f7e4931f8dc63b1cde9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5339885
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92781}