[cleanup][wasm] Add inlining positions to NativeModule::AddCode
AddCode() is only called for wrappers, so the inlining_positions
is always empty. Still, as all use the WasmCompilationResult, it's just
as easy to pass through the vector from the compilation result as to
create a dummy vector.

Change-Id: I3288c4239d3d9788664434b92186349520845b07
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5367023
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92799}