[wasm] Remove tier-up of js-to-js wrapper
The js-to-js wrapper gets used when a JS function gets wrapped with the
WebAssembly.Function constructor and then called from JavaScript. This
scenario is, however, unlikely and most likely not performance critical.
Therefore, this CL removes the tier-up of the js-to-js wrapper. The CL
removes the code that triggers tier-up, and it removes the TurboFan code
to compile an optimized js-to-js wrapper.

R=thibaudm@chromium.org

Bug: 328918912
Change-Id: I747c6a17799903a3d00a6b3d49f1279abf6f98b8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5366645
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92820}