[turboshaft][x64] Add x64 instruction selector unittest (disabled)
This CL just copy-pastes Turbofan's
instruction-selector-x64-unittest.cc without any changes beyond
disabling it with `#if 0`.

This looks like a fairly useless commit, but now we can port the
instruction-selector-x64-unittest to Turboshaft, and reviewing
will be much easier (because we'll see the delta from Tubofan to Turboshaft, which we wouldn't see easily if we were to port this
file in the same CL that creates it).

Bug: v8:12783
Change-Id: I35ff05e3b47ec2259e51738a705caf90275df433
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5368860
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92845}