[wasm] Refactor APIs between internal and external functions
Refactor the APIs for linking between internal and external functions.

The main motivation is the ReplaceWrapper function in which we know that
both the internal and the external function already exists, so there is
no point in calling a "GetOrCreate" accessor.

This change ensures that we don't accidentally create such objects while
replaceing wrapper code, and it also allows us to unhandlify a bunch of
accessors.

Drive-by: Avoid creating a HandleScope in the common cases in
WasmGenericWasmToJSObject.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: I9bb39e2cc4b0c5a73d4d1941acc19c209c810387
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5361783
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92848}