[arm64][cfi] Add "cc" to clobber list in authentication inline asm
When authenticating in inline assembly, we do a comparison to check if
the authentication was successful. This means we write to the
PSTATE.NZCV flags and so need to tell the compiler it cannot insert this
assembly between a comparison and a branch.

Add "cc" to the clobber list to fix this.

As a quick cleanup, add a comment clarifying why LR is restored, as it's
not obvious.

Bug: v8:10026
Change-Id: I62ad7bab72cb686cde71be9d657942c7930b86a4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5372731
Auto-Submit: Pierre Langlois <pierre.langlois@arm.com>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92869}