[wasm-gc][liftoff] Fix off-by-one in RefAsNonNull
When we implement `ref.as_non_null` as a fake load from the object
(which would trap if the object is null, handled by the trap handler),
we should take into account that the object base address is tagged,
so the "word after the map word" is at offset +3, not +4. Otherwise,
in the special case where the object is a non-null zero-fields struct
that was allocated right before the end of a page, we'd get an invalid
memory access and hence erroneously consider the struct to be `null`.

Fixed: 329145803
Change-Id: Ib8ce2668b767de9de6bf4632b022c71a3b8bd027
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5378140
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92879}