[profiler] Avoid data race in test by increasing sampling interval
There seems to be a data race in the CPU profiler when two samples get
collected at the same time. This happens in tests where samples get
collected explicitly through a test API, in addition to the samples that
get collected periodically by the profiler.

With this CL we increase the sampling interval of the profiler, so that
the test finishes before the profiler wants to collect its sample.

R=cbruni@chromium.org

Bug: v8:14549
Change-Id: Ie7678961e6b10a5250b2460b79a046f9c6b864ea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5377871
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92887}