[inspector] Allow custom formatters to return `null` from `body()`.
As per (unofficial) custom formatter specification[1], it is perfectly
valid to return `null` from the `body()` function, and the developer
tools should fall back to using the default body formatter in those
cases. This CL removes the check from the inspector, that the result
of calls to `body()` has to be an `Array`.

[1]: https://firefox-source-docs.mozilla.org/devtools-user/custom_formatters/index.html#custom-formatter-structure

Bug: b:329400119
Change-Id: I901116ef79bc169163f83d0db4aabe182f7ae4f2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5378291
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92930}