[wasm] Use TrustedFixedAddressArray for imported function targets
This CL introduces trusted versions of the integer arrays (based on the
existing TrustedByteArray) and uses a TrustedFixedAddressArray to store
the call targets of imported functions.

This closes a gap in the V8 sandbox which allowed attackers to easily
manipulate call targets stored on the JS heap.

R=ahaas@chromium.org

Bug: v8:14564, 40927374
Change-Id: Ifc477274c5652f3b87287aa89d0d325047c0c484
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5380602
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92936}