[turboshaft] Normalize graph before instruction selection
The instruction selector expects constants to be on the right-hand
side of commutative binops. Most of the time,
MachineOptimizationReducer is enough to maintain this invariant.
However, CodeEliminationAndSimplificationPhase runs last in the
pipeline, and can perform simplifications that lead to constants being
on the left-hand side (for instance, when removing a Phi with a single
input, it's possible that this input was a Constant, so it should go
on the right side, but before it looked like a Phi and thus could
have been on the left side).

This CL introduced InstructionSelectionNormalizationReducer, which
runs as part of CodeEliminationAndSimplificationPhase and puts
constants on the right side of commutative binops.
Additionally, it replaces multiplications by small constants by
shifts.

Fixed: 330638807
Bug: v8:12783
Change-Id: Iac95f8e49f683ea5e5f502c96e76f442831267ff
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5388437
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93005}