[ro-promotion][sandbox] Fix promotion of code table entries
After read-only promotion, some Code objects are promoted to the RO
space. But the corresponding code pointer table (CPT) entries are not
in the read-only segment. Therefore, if a GC is triggered at this
point, they will be swept and this will leave such Code objects with
an invalid self indirect pointer offset. If heap verification (e.g.,
as part of the same GC) ever reaches such an object, it will fail.

This CL:

1. Adds the following invariant to the heap verifier: if an
   ExposedTrustedObject is in the RO space, its self indirect pointer
   entry must be in the read-only segment, and vice versa.
2. Fixes read-only promotion, so that code pointer table entries are
   correctly promoted and the above invariant is satisfied.
3. Triggers the verification of the mutable heap (as well as that of
   the read-only heap) after read-only promotion.
4. Modifies cctest/ContextSerializerCustomContext so that a GC is
   triggered after read-only promotion (which crashes without the
   fix in this CL).

Bug: b/330450848
Change-Id: I5525579339a085082a4dc935f33a9800e1b946e0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5383045
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93076}