[compiler] Allow implicit truncation for graphs created from turbofan
It would be nice to not have implicit truncations from Word64 to
Word32 in Turboshaft as they are inconsistently sometimes explicit and
sometimes implicit.

However, as the same was true for Turbofan it is difficult to enforce
it on Turboshaft graphs created from Turbofan.

To have this asserted in Turboshaft-only cases (wasm, unittests,
Maglev->Turboshaft), pass this information downwards and relax the
assertion again in ValidOpInputRep.

Note: The implicit truncation was disallowed in
https://crrev.com/c/5380531 .

Fixed: 330638812
Change-Id: Id67d5781dfac88fd37131c0f49eb3fafa6cdcf0f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5385337
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93022}