Fix missing float16 enabled check on the serializer
Due to the ArrayBuffer possibly being modified in the script,
the serialized value type can be changed to 'h' which means
a half-precision floating point.

Since float16 is an experimental feature and should not be
exposed without a flag, on reading the array buffer,
Split the code path float16 or not and add a conditional statement
that checks the float16 enabled.

Bug: 330922408
Change-Id: If7bff8812cc5b6e6450c257bbb6d0b2944d1ffce
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5394893
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Commit-Queue: Seokho Song <seokho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93125}