[runtime] Allow exceptions in StringEscapeQuotes
If the input string is close to the maximum string length, then the
escaped string can be too large. In that case the factory raises a
RangeError.
This CL changes the runtime function to account for the possibility of
this exception, and adds a regression test.

R=ishell@chromium.org

Bug: chromium:331074427
Change-Id: I7ad0c18e5b0beb149441a40aa14828a4acbd496e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5392585
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93042}