[sandbox] Introduce SharedFunctionInfoWrapper
This will be necessary when converting DeoptimizationData to a
ProtectedFixedArray, at which point it can only reference other trusted
objects (or Smis). As such, the current reference to the in-sandbox
SharedFunctionInfo must then be replaced with a reference to the SFI's
SharedFunctionInfoWrapper in trusted space.

Bug: chromium:329677666
Change-Id: I1e5889bc39f15a0ebbea0f057731e450275e38fb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5371949
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93050}