[liftoff] Add helpers for managing temp registers
This adds the TempRegisterScope and ScopedTempRegister classes.
The first is a scope which manages temporary registers that are reused
for different things in the course of generating code for an
instruction.
The latter is a scoped object which holds a temp registers and returns
it to the first class when it gets out of scope.

This will hopefully make it easier to write code which uses the same set
of temporaries for different values during code generation in Liftoff.

For now we only use it in CallIndirect which manages three temporaries.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: Id2d5d5f6d945ec7109029dd1c843fee08f1b6167
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5393445
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93068}