[maglev] Add FastContexts to escape analysis
That is, it allows the allocation of contexts to be elided if
it does not escape.

This CL:
- simplifies FastContext, since most fields are just undefined. This
allows to easily identify runtime values. FastContext also does not
use FastField(ValueNode*) anymore, which makes it possible to
delete this constructor in a follow up CL.
- updates DeepForEachInputImpl to also visit DeoptObjects.
- adds InputLocations in DeoptInfos for value nodes in a DeoptObject
- extends the lifetime of value nodes that are fields in a DeoptObject
- allocates inputs for these value nodes during regalloc
- create a CAPTURE_OBJECT for non escaping contexts to materialize
them during deoptimization.

Bug: 326091155

Change-Id: I58a446bb17b702cb3dd90564c6793d16bcb7d43e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5392586
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93105}