[maglev] Fix strict comparison HoleyFloat64 with constants
We assume that if we strict compare an untagged value to a constant,
then the comparison is false.

This is not true with HoleyFloat64, since they can compare
equal to the undefined value.

This adds HoleyFloat64IsHole node and the optimization to use
BranchIfFloat64IsHole.

Fixed: 331711028
Change-Id: I9e2ccbdeda1f12795406a1256b5a76e526453d27
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5401864
Reviewed-by: Patrick Thier <pthier@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93102}