[sandbox] Introduce the sandbox testing mode
The sandbox-testing mode is the generic mode for testing the sandbox's
security properties. It is used specifically for demonstrating and
validating sandbox bypasses. Currently, the requirements for a sandbox
bypass are fairly strict: there must be a write to a specific page
allocated at a random address outside of the sandbox at startup. In the
future, it's expected that these requirements will eventually be relaxed
to simply "memory corruption outside the sandbox".

Bug: v8:12878
Change-Id: Ifae611e3bb291c98d54ca26b26c1336b3e43044e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5412311
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93153}