[liftoff] Optimize static index for indirect call
If the index for an indirect call is statically known we can generate
much leaner code by avoiding the computation of the dispatch table
offset.

Also in the case of a dynamic index this CL makes us emit one less
addition, and instead adds the kEntriesOffset together with the kRefBias
or kTargetBias later.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: Id92355a162fd260d517e172555c3d2801b3c2a94
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5401777
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93114}