[string + string wrapper] Fix the case when we subclass String
If we subclass String and the subclass has a custom @@toPrimitive
function, we didn't detect it and ignore it (in all tiers).

This CL invalidates the "string wrapper toPrimitive" protector also
when we detect that something subclasses String (whether or not it
has a custom toPrimitive) and when calling Reflect.construct with
a String and a non-String new.target.

Bug: v8:14688, v8:12199
Change-Id: I65668e25c4dc7bd394cb80b569907a90c701f669
Fixed: v8:14688
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5401788
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93118}