[frames] Merge frame singletons into a union
Remove the concept of individual frame singletons in the stack iterator,
replacing it with a union over the different frame types. This reduces
the stack size of the stack iterator (which is important for profiling
stack iterators that have a small stack in the signal handler).

This requires in-place initialising the frame when setting it, which in
turn requires updating code which thinks it can call SingletonFor
without mutating the current frame. So, remove SingletonFor entirely,
making it clear that this is a frame initialisation, and clean up the
code that (mis)used it.

Change-Id: I26b010415334ec6ae8c70ed10e828a19547d2470
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5412661
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93120}