[shared-struct] Notify cv waiters while holding the queue lock
When notifying a JS condition variable, currently we dequeue all the
waiters, release the queue lock, and then notify the waiters. This can
be an issue if when trying to notify a waiter that is being destroyed
in another thread. This will be a bigger issue for async waiters but
can happen with sync waiters that have timed out around the same time
a notify is called in another thread.

Drive-by: Scope the WaiterQueueNode lifetime in
JSAtomicsCondition::WaitFor so that the waiter is deleted before trying
to request the JS mutex.

Bug: v8:12547
Change-Id: I8d9a46082c8c99e7534eef8970d1be93ada40416
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5410368
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Commit-Queue: Luis Fernando Pardo Sixtos <lpardosixtos@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#93178}