[turbolizer] Increase the Z position of the source position marker
The source position marker can be located on the other line.
Because the marker can overlap with the next line, it can be hard to click.

Since the positions are all static, we cannot use z-index.

Hence, increase the Z position to lift up the clickable area to resolve this problem simple way.

Bug: v8:7327
Change-Id: Ifffbb6487c93428e90eb1a3bbeea68ffbdef8d02
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5422289
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93379}