[cppgc] Sweep custom spaces first
Blink's HeapVectors are the most likely to result in large objects that
require finalization on the main thread. This CL sweeps the custom
spaces first so that the large finalizations can occur while concurrent
sweeping is still active, rather than after it is done.

Change-Id: I520ab4117c6eb1d8876df86ec292da2101a5ca60
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5418934
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93161}