[error] Use message passed to constructor for Error.stack message
Even though Error.stack is now a true JS accessor, the content itself
is cemented after the first time the Error.stack getter is called.

This means modifying the 'message' property of an Error object may or
may not have any affect on the stack trace depending on whether
Error.stack was accessed previously.

This can be particularly confusing in the DevTools console or Node
REPL, where a preview is built as the user types, and Error.stack is
accessed as a side-effect.

This CL proposes to make this more consistent: When V8 constructs
a new Error object, we stash the original message in a private
symbol similar to Error.stack. When we serialize Error.stack, we
use the original message from the private symbol, rather than the
current value of the 'message' property.

If no private symbol is present along the prototype chain,
we use the 'message property'.

This CL does not change Error.captureStackTrace. For captureStackTrace
it depends if the custom object subclasses "Error" and how the super
constructor is called (only the Error constructor installs the private
message symbol).

Bug: 327467399
Change-Id: I1ea92c96b96fa95bfc4fb8dcf1571bd647b3c68f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5378709
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93164}