[heap] --no-concurrent-marking should also affect cppgc
This CL fixes a crash in chromium when running with
"--no-concurrent-marking --no-parallel-marking". When disabling
parallel marking we do not finish concurrent marking. This is not
an issue for V8 because disabling parallel marking also disables
concurrent marking through flag implications. This CL solves this
issue by disabling concurrent marking for cppgc as well when
"--no-concurrent-marking" is given as an argument.

Alternatively we could also finish concurrent marking for cppgc.
However, we want flags like that anyways to apply to both GCs and
we already have such an implication for "--no-incremental-marking".

Bug: chromium:40278463
Change-Id: Ib29d39fa98fdb6f0efdb2f0f2d32a9140524ec0c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5424090
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93190}