[turboshaft][wasm-imported-strings] Annotate that some builtins can end with trap
The changed imported-string builtins can trap (see wasm.tq). However,
this was not annotated in the builtins. Therefore, these operations
might be considered as eliminatable, and if the result is unused, the
operation can be dropped even if it would have trapped.

Fixed: 40258436
Change-Id: I393ef66f9af213da1c59a2a7a87f3f33573138de
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5419422
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93171}