[heap] move MemoryChunk::Metadata accessor to inl.h
In preparation for sandboxifying the metadata_ pointer. In the next CL,
we need the definition of the Metadata class inside the
MemoryChunk::Metadata accessor, so move it to -inl.h. This in turn
requires a few other -inl.h files to be added.

Change-Id: Ibc36e5489fd2f42895a55f7ec4e920ade241f13d
Bug: chromium:40849120
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5412132
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#93223}