[liftoff] Bump constants for reducing tiering budget
Recent numbers seen in experiments seem to suggest that we trigger
tier-up of often called functions very late. And indeed for very small
functions we only reduce the tiering budget by a small amount. The
actual time spent in the function is actually larger though, because we
do not account for the time spent in the WasmLiftoffFrameSetup builtin,
and for the code executed for the actual tier-up check.

This CL bumps the constant for the latter from 1 to 20 (which is closer
to the actual code size on x64), and introduces a new constant for the
former according to the size of that builtin.

R=dlehmann@chromium.org

Change-Id: I304c9de63bcc8db1be0150e1c7b6ee52bd47430d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5424734
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93225}