[maps] Always apply prototype transitions first
I.e., treat prototype transitions similar to elements kind
transitions. We move prototype transitions to the start of the map
tree, before ElementsKind transitions, using the MapUpdater.

This allows us to finally not generalize all fields anymore on the
target map of UpdatePrototype. The additional benefit is that it
prevents us from collecting the map tree before a prototype
transition.

Say we have the following code:

    var x = new foo()
    x.a = 1.1
    x.__proto__ = bar.prototype


Before this CL we produce the following transition tree, where double
arrows are strong and single arrows are weak pointers:

    foo {} <==--> foo {a} ----> bar {a} <==== x
                  ~~~~~~~

Assuming x to be a live object, then the underlined subtree is still
only weakly reachable from x. In other words it can be collected and
if we later create a second such object, the existing `bar {a}` map is
not reachable anymore. We re-create the whole subtree and the new
object will have a fresh (but identical) map.

The problems are:
1. Accidental polymorphism.
2. Unnecessarily duplicated maps.
3. Prototype transitions must generalize all fields
   (because we have no back-pointers and thus field updates cannot be
    properly back-propagated).

The solution here is to move prototype transitions to the beginning of
the map tree:

    foo {} <==--> foo {a}
    ∧
    `==--> bar {} <==--> bar {a} <==== x

And use the full MapUpdater to get from `foo {a}` to `bar {a}`
This fixes all issues. Generalization do not happens instantly, but
rather during the MapUpdate on UpdatePrototype.

Bug: chromium:332914927

Change-Id: I947e45448a9de1b14f4f8c0c4e1081910027390a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5300551
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93189}