[wasm] Call on-heap import wrappers via a trampoline
This allows us to remove the `code` field from WasmInternalFunction and
always call the stored `call_target`.
The only situation where we still needed the fallback to call via the
code pointer was for on-heap wasm-to-js wrappers. Those always receive
the WasmApiFunctionRef as first parameter, so we can instead store the
code pointer there and call a builtin which loads the instruction start
from the code pointer and then tail-calls (jumps) there.

This generates smaller code for every call_ref by avoiding the check for
`call_target == 0` and the fallback to load the instruction start from
the code pointer table.

R=ahaas@chromium.org
CC=manoskouk@chromium.org

Bug: v8:14564
Change-Id: I40ea7d9b409dc654bd90c3360c0bf064cf4978ab
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5419396
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93191}