[sandbox] Introduce ExternalPointerTable::ManagedResource
This is a (malloced) object located outside of the V8 heap whose
lifetime is tied to something inside the V8 heap. Examples include the
objects referenced via Managed objects or ArrayBufferExtensions. This
class is used to make this relationship explicit, which is necessary for
the sandbox: as these objects are referenced through the external
pointer table (EPT), it must be ensured that they are never deallocated
before their EPT entry is freed, or if they are, the entry must be
zapped/invalidated. Failure to do so results in a dangling EPT entry
that can lead to a use-after-free outside of the sandbox.

For now, the ManagedResource does the latter: it keeps track of its
entry in the EPT and zaps the entry when the resource itself is
deallocated. However, in the future it might be more efficient to tie
the lifetime of the external resource directly to the table entry and
only free the resource when the table entry is freed. That, however,
likely first requires minor GC support for the EPT so that resources are
not kept around longer than necessary.

Bug: chromium:331237575
Change-Id: I13c4c9dcac101aa1f6d10b902aac9260f68fdbda
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5425025
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93228}