[turboshaft][wasm-imported-strings] Annotate that some more functions can end with trap
The changed stringref runtime functions can trap (see runtime-wasm.cc).
Similarly as in https://crrev.com/c/5419422:
The ability to trap was not annotated in these functions. Therefore, these operations
might be considered as eliminatable, and if the result is unused, the operation can be
dropped even if it would have trapped.

Fixed: 330580823
Change-Id: I4f2adec192b918f8ca32f2fcee326d8c1b6eba92
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5426675
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93196}