[maglev] Fix DeepForEachInputAndDeoptObject to consider nested captured objects
The function passed as parameter to DeepForDeoptObject tree walker needs
to consider nested escaped objects. We use a recursive lambda for that.

In the test case, for instance, we should not try to apply argument `f`
(the node marker) to a captured context (InlinedAllocation) that is
the extension value of another captured context that will be
materialized when/if we deopt.

Fixed: 332947120
Bug: 326091155
Change-Id: Ibab8bb758cc7ce704cfeee4200f6fcab0eaf4728
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5431937
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93229}